requirements.txt
Flask==2.1.4
Flask-SQLAlchemy==3.0.0
python-dotenv==0.19.1
waitress==2.1.0
pyodbc==4.0.32
